backend/fastapi-graphql/pyproject.toml
[tool.poetry]
name = "fastapi-graphql"
version = "0.1.0"
description = ""
authors = ["Wickes1 <27881570+wickes1@users.noreply.github.com>"]
readme = "README.md"
packages = [{ include = "fastapi_graphql" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
sqlmodel = "^0.0.16"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
strawberry-graphql = { extras = ["fastapi"], version = "^0.224.1" }
asyncpg = "^0.29.0"

[tool.poetry.scripts]
start = "fastapi_graphql.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
